[css-variables] Clone style when there are no variable references
When resolving the CSS Custom Properties inheritance, we recalculate all
the styles of nodes in the hierarchy. Following a similar approach to
the one used for 'independent properties', we can just clone the node's
parent style and propagate just the InheritedVariables data structure.

We'll only recalc the styles if a node has any variable reference,
defined as :var(--*), to a custom property.

Bug: 1056209
Change-Id: I81081e11cd69f06e69a6030b97fabce3a50630b3